requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.7,<3.0
langgraph>=0.2.30,<1.0
openai>=1.40,<2.0
requests>=2.31,<3.0
beautifulsoup4>=4.12,<5.0
duckduckgo-search>=6.2,<9.0
python-dotenv>=1.0,<2.0
sentence-transformers>=3.0,<6.0
faiss-cpu>=1.8,<2.0
numpy>=1.26,<3.0
streamlit>=1.34,<2.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
